+++ src/gui/style.rs
@@ -25,6 +25,7 @@ impl ColorExt for Color {
 
 pub struct Theme {
     background: Color,
+    background_secondary: Color,
     field: Color,
     text: Color,
     text_inverted: Color,
@@ -52,6 +53,7 @@ impl From<config::Theme> for Theme {
         match source {
             config::Theme::Light => Self {
                 background: Color::WHITE,
+                background_secondary: rgb8!(210, 210, 210),
                 field: rgb8!(230, 230, 230),
                 text: Color::BLACK,
                 text_inverted: Color::WHITE,
@@ -69,6 +71,7 @@ impl From<config::Theme> for Theme {
             },
             config::Theme::Dark => Self {
                 background: rgb8!(41, 41, 41),
+                background_secondary: rgb8!(100, 100, 100),
                 field: rgb8!(74, 74, 74),
                 text: Color::WHITE,
                 text_inverted: Color::BLACK,
@@ -251,7 +254,8 @@ impl container::StyleSheet for Theme {
         container::Appearance {
             background: match style {
                 Self::Style::Wrapper => Color::TRANSPARENT.into(),
-                Self::Style::Secondary | Self::Style::GameListEntry => self.field.alpha(0.15).into(),
+                Self::Style::Secondary => self.background_secondary.into(),
+                Self::Style::GameListEntry => self.field.alpha(0.15).into(),
                 Self::Style::ModalBackground | Self::Style::Notification | Self::Style::Tooltip => self.field.into(),
                 Self::Style::DisabledBackup => self.disabled.into(),
                 Self::Style::BadgeActivated => self.negative.into(),